+++ packages/core/src/RenderingEngine/Viewport.ts
@@ -10,6 +10,7 @@ import Events from '../enums/Events';
 import ViewportType from '../enums/ViewportType';
 import renderingEngineCache from './renderingEngineCache';
 import { triggerEvent, planar, isImageActor } from '../utilities';
+import hasNaNValues from '../utilities/hasNaNValues';
 import { RENDERING_DEFAULTS } from '../constants';
 import type {
   ICamera,
@@ -173,24 +174,6 @@ class Viewport implements IViewport {
     }
   }
 
-  protected applyFlipTx = (worldPos: Point3): Point3 => {
-    const actorEntry = this.getDefaultActor();
-
-    if (!actorEntry || !actorEntry.actor) {
-      return worldPos;
-    }
-
-    const actor = actorEntry.actor;
-    const mat = actor.getMatrix();
-
-    const newPos = vec3.create();
-    const matT = mat4.create();
-    mat4.transpose(matT, mat);
-    vec3.transformMat4(newPos, worldPos, matT);
-
-    return [newPos[0], newPos[1], newPos[2]];
-  };
-
   /**
    * Flip the viewport on horizontal or vertical axis, this method
    * works with vtk-js backed rendering pipeline.
@@ -206,105 +189,115 @@ class Viewport implements IViewport {
       return;
     }
 
-    let flipH = false;
-    let flipV = false;
+    const { viewPlaneNormal, viewUp, focalPoint, position } = this.getCamera();
 
-    if (
-      typeof flipHorizontal !== 'undefined' &&
-      ((flipHorizontal && !this.flipHorizontal) ||
-        (!flipHorizontal && this.flipHorizontal))
-    ) {
-      flipH = true;
+    const viewRight = vec3.create();
+    vec3.cross(viewRight, viewPlaneNormal, viewUp);
+
+    let viewUpToSet = vec3.create();
+    vec3.copy(viewUpToSet, viewUp);
+
+    let viewPlaneNormalToSet = vec3.create();
+    viewPlaneNormalToSet = vec3.negate(viewPlaneNormalToSet, viewPlaneNormal);
+
+    // for both flip horizontal and vertical we need to move the camera to the
+    // other side of the image
+    const distance = vec3.distance(position, focalPoint);
+
+    // If the pan has been applied, we need to be able
+    // apply the pan back
+    const resetFocalPoint = vec3.create();
+    const dimensions = imageData.getDimensions();
+    const middleIJK = dimensions.map((d) => Math.floor(d / 2));
+
+    const idx = [middleIJK[0], middleIJK[1], middleIJK[2]];
+    imageData.indexToWorld(idx, resetFocalPoint);
+
+    // what is the difference right now between the rested focal point and
+    // the current focal point
+    // Todo: this needs to be retrieved from the function that considers maintainFrame
+    // just now trying it on stack Viewport
+    const panDir = vec3.create();
+    vec3.subtract(panDir, focalPoint, resetFocalPoint);
+
+    const panValue = vec3.length(panDir);
+
+    const getPanDir = (mirrorVec) => {
+      const panDirMirror = vec3.create();
+      vec3.scale(panDirMirror, mirrorVec, 2 * vec3.dot(panDir, mirrorVec));
+      vec3.subtract(panDirMirror, panDirMirror, panDir);
+      vec3.normalize(panDirMirror, panDirMirror);
+
+      return panDirMirror;
+    };
+
+    // Flipping horizontal mean that the camera should move
+    // to the other side of the image but looking at the
+    // same direction and same focal point
+    if (flipHorizontal) {
+      // we need to apply the pan value to the new focal point but in the direction
+      // that is mirrored on the viewUp for the flip horizontal and
+      // viewRight for the flip vertical
+      const newFocalPoint = vec3.create();
+
+      // mirror the pan direction based on the viewUp
+      const panDirMirror = getPanDir(viewUpToSet);
+
+      // move focal point from the resetFocalPoint to the newFocalPoint
+      // based on the panDirMirror and panValue
+      vec3.scaleAndAdd(newFocalPoint, resetFocalPoint, panDirMirror, panValue);
+
+      // move the camera position also the same way as the focal point
+      const newPosition = vec3.create();
+
+      vec3.scaleAndAdd(
+        newPosition,
+        newFocalPoint,
+        viewPlaneNormalToSet,
+        distance
+      );
+
+      this.setCamera({
+        viewPlaneNormal: viewPlaneNormalToSet as Point3,
+        position: newPosition as Point3,
+        focalPoint: newFocalPoint as Point3,
+      });
+
+      this.flipHorizontal = !this.flipHorizontal;
     }
 
-    if (
-      typeof flipVertical !== 'undefined' &&
-      ((flipVertical && !this.flipVertical) ||
-        (!flipVertical && this.flipVertical))
-    ) {
-      flipV = true;
+    // Flipping vertical mean that the camera should negate the view up
+    // and also move to the other side of the image but looking at the
+    if (flipVertical) {
+      viewUpToSet = vec3.negate(viewUpToSet, viewUp);
+
+      // we need to apply the pan value to the new focal point but in the direction
+      const panDirMirror = getPanDir(viewRight);
+
+      const newFocalPoint = vec3.create();
+
+      vec3.scaleAndAdd(newFocalPoint, resetFocalPoint, panDirMirror, panValue);
+
+      const newPosition = vec3.create();
+
+      vec3.scaleAndAdd(
+        newPosition,
+        newFocalPoint,
+        viewPlaneNormalToSet,
+        distance
+      );
+
+      this.setCamera({
+        focalPoint: newFocalPoint as Point3,
+        viewPlaneNormal: viewPlaneNormalToSet as Point3,
+        viewUp: viewUpToSet as Point3,
+        position: newPosition as Point3,
+      });
+
+      this.flipVertical = !this.flipVertical;
     }
 
-    if (!flipH && !flipV) {
-      return;
-    }
-
-    // In Cornerstone gpu rendering pipeline, the images are positioned
-    // in the space according to their origin, and direction (even StackViewport
-    // with one slice only). In order to flip the images, we need to flip them
-    // around their center axis (either horizontal or vertical). Since the images
-    // are positioned in the space according to their origin and direction, for a
-    // proper scaling (flipping), they should be transformed to the origin and
-    // then flipped. The following code does this transformation.
-
-    const origin = imageData.getOrigin();
-    const direction = imageData.getDirection();
-    const spacing = imageData.getSpacing();
-    const size = imageData.getDimensions();
-
-    const iVector = direction.slice(0, 3);
-    const jVector = direction.slice(3, 6);
-    const kVector = direction.slice(6, 9);
-
-    // finding the center of the image
-    const center = vec3.create();
-    vec3.scaleAndAdd(center, origin, iVector, (size[0] / 2.0) * spacing[0]);
-    vec3.scaleAndAdd(center, center, jVector, (size[1] / 2.0) * spacing[1]);
-    vec3.scaleAndAdd(center, center, kVector, (size[2] / 2.0) * spacing[2]);
-
-    let flipHTx, flipVTx;
-
-    const transformToOriginTx = vtkMatrixBuilder
-      .buildFromRadian()
-      .identity()
-      .translate(center[0], center[1], center[2])
-      .rotateFromDirections(jVector, [0, 1, 0])
-      .rotateFromDirections(iVector, [1, 0, 0]);
-
-    const transformBackFromOriginTx = vtkMatrixBuilder
-      .buildFromRadian()
-      .identity()
-      .rotateFromDirections([1, 0, 0], iVector)
-      .rotateFromDirections([0, 1, 0], jVector)
-      .translate(-center[0], -center[1], -center[2]);
-
-    if (flipH) {
-      this.flipHorizontal = flipHorizontal;
-      flipHTx = vtkMatrixBuilder
-        .buildFromRadian()
-        .multiply(transformToOriginTx.getMatrix())
-        .scale(-1, 1, 1)
-        .multiply(transformBackFromOriginTx.getMatrix());
-    }
-
-    if (flipV) {
-      this.flipVertical = flipVertical;
-      flipVTx = vtkMatrixBuilder
-        .buildFromRadian()
-        .multiply(transformToOriginTx.getMatrix())
-        .scale(1, -1, 1)
-        .multiply(transformBackFromOriginTx.getMatrix());
-    }
-
-    const actorEntries = this.getActors();
-
-    actorEntries.forEach((actorEntry) => {
-      const actor = actorEntry.actor;
-
-      const mat = actor.getUserMatrix();
-
-      if (flipHTx) {
-        mat4.multiply(mat, mat, flipHTx.getMatrix());
-      }
-
-      if (flipVTx) {
-        mat4.multiply(mat, mat, flipVTx.getMatrix());
-      }
-
-      actor.setUserMatrix(mat);
-    });
-
-    this.getRenderingEngine().render();
+    this.render();
   }
 
   private getDefaultImageData(): any {
@@ -545,9 +538,19 @@ class Viewport implements IViewport {
   protected resetCamera(
     resetPan = true,
     resetZoom = true,
+    resetToCenter = true,
     storeAsInitialCamera = true
   ): boolean {
     const renderer = this.getRenderer();
+
+    // fix the flip right away, since we rely on the viewPlaneNormal and
+    // viewUp for later. Basically, we need to flip back if flipHorizontal
+    // is true or flipVertical is true
+    this.setCamera({
+      flipHorizontal: false,
+      flipVertical: false,
+    });
+
     const previousCamera = _cloneDeep(this.getCamera());
 
     const bounds = renderer.computeVisiblePropBounds();
@@ -573,7 +576,6 @@ class Viewport implements IViewport {
     // Reset the perspective zoom factors, otherwise subsequent zooms will cause
     // the view angle to become very small and cause bad depth sorting.
     // todo: parallel projection only
-    activeCamera.setViewAngle(90.0);
 
     focalPoint[0] = (bounds[0] + bounds[1]) / 2.0;
     focalPoint[1] = (bounds[2] + bounds[3]) / 2.0;
@@ -624,53 +626,51 @@ class Viewport implements IViewport {
     radius = Math.sqrt(radius) * 0.5;
 
     const distance = 1.1 * radius;
-    // const distance = radius / Math.sin(angle * 0.5)
 
-    // check view-up vector against view plane normal
-    if (Math.abs(vtkMath.dot(viewUp, viewPlaneNormal)) > 0.999) {
-      activeCamera.setViewUp(-viewUp[2], viewUp[0], viewUp[1]);
-    }
+    const viewUpToSet: Point3 =
+      Math.abs(vtkMath.dot(viewUp, viewPlaneNormal)) > 0.999
+        ? [-viewUp[2], viewUp[0], viewUp[1]]
+        : viewUp;
 
-    const focalPointToSet = resetPan ? focalPoint : previousCamera.focalPoint;
-
-    activeCamera.setFocalPoint(
-      focalPointToSet[0],
-      focalPointToSet[1],
-      focalPointToSet[2]
+    const focalPointToSet = this._getFocalPointForResetCamera(
+      focalPoint,
+      previousCamera,
+      { resetPan, resetToCenter }
     );
-    activeCamera.setPosition(
+
+    const positionToSet: Point3 = [
       focalPointToSet[0] + distance * viewPlaneNormal[0],
       focalPointToSet[1] + distance * viewPlaneNormal[1],
-      focalPointToSet[2] + distance * viewPlaneNormal[2]
-    );
+      focalPointToSet[2] + distance * viewPlaneNormal[2],
+    ];
 
     renderer.resetCameraClippingRange(bounds);
 
-    if (resetZoom) {
-      activeCamera.setParallelScale(parallelScale);
-    }
+    const clippingRangeToUse: Point2 = [
+      -RENDERING_DEFAULTS.MAXIMUM_RAY_DISTANCE,
+      RENDERING_DEFAULTS.MAXIMUM_RAY_DISTANCE,
+    ];
 
-    // update reasonable world to physical values
     activeCamera.setPhysicalScale(radius);
-
-    // TODO: The PhysicalXXX stuff are used for VR only, do we need this?
     activeCamera.setPhysicalTranslation(
       -focalPointToSet[0],
       -focalPointToSet[1],
       -focalPointToSet[2]
     );
 
-    activeCamera.setClippingRange(
-      -RENDERING_DEFAULTS.MAXIMUM_RAY_DISTANCE,
-      RENDERING_DEFAULTS.MAXIMUM_RAY_DISTANCE
-    );
+    this.setCamera({
+      parallelScale: resetZoom ? parallelScale : previousCamera.parallelScale,
+      focalPoint: focalPointToSet,
+      position: positionToSet,
+      viewAngle: 90,
+      viewUp: viewUpToSet,
+      clippingRange: clippingRangeToUse,
+    });
 
-    if (this.flipHorizontal || this.flipVertical) {
-      this.flip({ flipHorizontal: false, flipVertical: false });
-    }
+    const modifiedCamera = _cloneDeep(this.getCamera());
 
     if (storeAsInitialCamera) {
-      this.setInitialCamera(this.getCamera());
+      this.setInitialCamera(modifiedCamera);
     }
 
     const RESET_CAMERA_EVENT = {
@@ -682,10 +682,7 @@ class Viewport implements IViewport {
     // and do the right thing.
     renderer.invokeEvent(RESET_CAMERA_EVENT);
 
-    this.triggerCameraModifiedEventIfNecessary(
-      previousCamera,
-      this.getCamera()
-    );
+    this.triggerCameraModifiedEventIfNecessary(previousCamera, modifiedCamera);
 
     return true;
   }
@@ -859,8 +856,8 @@ class Viewport implements IViewport {
     return {
       viewUp: <Point3>vtkCamera.getViewUp(),
       viewPlaneNormal: <Point3>vtkCamera.getViewPlaneNormal(),
-      position: <Point3>this.applyFlipTx(vtkCamera.getPosition() as Point3),
-      focalPoint: <Point3>this.applyFlipTx(vtkCamera.getFocalPoint() as Point3),
+      position: <Point3>vtkCamera.getPosition(),
+      focalPoint: <Point3>vtkCamera.getFocalPoint(),
       parallelProjection: vtkCamera.getParallelProjection(),
       parallelScale: vtkCamera.getParallelScale(),
       viewAngle: vtkCamera.getViewAngle(),
@@ -891,10 +888,34 @@ class Viewport implements IViewport {
       viewAngle,
       flipHorizontal,
       flipVertical,
+      clippingRange,
     } = cameraInterface;
 
-    if (flipHorizontal !== undefined || flipVertical !== undefined) {
-      this.flip({ flipHorizontal, flipVertical });
+    // Note: Flip camera should be two separate calls since
+    // for flip, we need to flip the viewportNormal, and if
+    // flipHorizontal, and flipVertical are both true, that would
+    // the logic would be incorrect. So instead, we handle flip Horizontal
+    // and flipVertical separately.
+    if (flipHorizontal !== undefined) {
+      // flip if not flipped but requested to flip OR if flipped but requested to
+      // not flip
+      const flipH =
+        (flipHorizontal && !this.flipHorizontal) ||
+        (!flipHorizontal && this.flipHorizontal);
+
+      if (flipH) {
+        this.flip({ flipHorizontal: flipH });
+      }
+    }
+
+    if (flipVertical !== undefined) {
+      const flipV =
+        (flipVertical && !this.flipVertical) ||
+        (!flipVertical && this.flipVertical);
+
+      if (flipV) {
+        this.flip({ flipVertical: flipV });
+      }
     }
 
     if (viewUp !== undefined) {
@@ -910,11 +931,11 @@ class Viewport implements IViewport {
     }
 
     if (position !== undefined) {
-      vtkCamera.setPosition(...this.applyFlipTx(position));
+      vtkCamera.setPosition(...position);
     }
 
     if (focalPoint !== undefined) {
-      vtkCamera.setFocalPoint(...this.applyFlipTx(focalPoint));
+      vtkCamera.setFocalPoint(...focalPoint);
     }
 
     if (parallelScale !== undefined) {
@@ -923,6 +944,10 @@ class Viewport implements IViewport {
 
     if (viewAngle !== undefined) {
       vtkCamera.setViewAngle(viewAngle);
+    }
+
+    if (clippingRange !== undefined) {
+      vtkCamera.setClippingRange(clippingRange);
     }
 
     // update clippingPlanes if volume viewports
@@ -1099,6 +1124,62 @@ class Viewport implements IViewport {
       [bounds[1], bounds[3], bounds[4]],
       [bounds[1], bounds[3], bounds[5]],
     ];
+  }
+
+  _getFocalPointForResetCamera(
+    centeredFocalPoint: Point3,
+    previousCamera: ICamera,
+    { resetPan, resetToCenter }: { resetPan: boolean; resetToCenter: boolean }
+  ): Point3 {
+    if (resetToCenter && resetPan) {
+      return centeredFocalPoint;
+    }
+
+    if (resetToCenter && !resetPan) {
+      return hasNaNValues(previousCamera.focalPoint)
+        ? centeredFocalPoint
+        : previousCamera.focalPoint;
+    }
+
+    if (!resetToCenter && resetPan) {
+      // this is an interesting case that means the reset camera should not
+      // change the slice (default behavior is to go to the center of the
+      // image), and rather just reset the pan on the slice that is currently
+      // being viewed
+      const oldCamera = previousCamera;
+      const oldFocalPoint = oldCamera.focalPoint;
+      const oldViewPlaneNormal = oldCamera.viewPlaneNormal;
+
+      const vectorFromOldFocalPointToCenteredFocalPoint = vec3.create();
+      vec3.subtract(
+        vectorFromOldFocalPointToCenteredFocalPoint,
+        centeredFocalPoint,
+        oldFocalPoint
+      );
+
+      const distanceFromOldFocalPointToCenteredFocalPoint = vec3.dot(
+        vectorFromOldFocalPointToCenteredFocalPoint,
+        oldViewPlaneNormal
+      );
+
+      const newFocalPoint = vec3.create();
+      vec3.scaleAndAdd(
+        newFocalPoint,
+        centeredFocalPoint,
+        oldViewPlaneNormal,
+        -1 * distanceFromOldFocalPointToCenteredFocalPoint
+      );
+
+      return [newFocalPoint[0], newFocalPoint[1], newFocalPoint[2]];
+    }
+
+    if (!resetPan && !resetToCenter) {
+      // this means the reset camera should not change the slice and should not
+      // touch the pan either.
+      return hasNaNValues(previousCamera.focalPoint)
+        ? centeredFocalPoint
+        : previousCamera.focalPoint;
+    }
   }
 
   /**