+++ packages/core/src/constants/epsilon.ts
@@ -1,0 +1,3 @@
+const EPSILON = 1e-3;
+
+export default EPSILON;