+++ packages/core/src/types/ICamera.ts
@@ -1,4 +1,5 @@
 import Point3 from './Point3';
+import Point2 from './Point2';
 
 /**
  * Camera Interface. See {@link https://kitware.github.io/vtk-examples/site/VTKBook/03Chapter3/#35-cameras} if you
@@ -28,6 +29,8 @@ interface ICamera {
   flipHorizontal?: boolean;
   /** flip Vertical */
   flipVertical?: boolean;
+  /** clipping range */
+  clippingRange?: Point2;
 }
 
 export default ICamera;