+++ packages/docs/CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.26](https://github.com/cornerstonejs/cornerstone3D-beta/compare/docs@0.6.25...docs@0.6.26) (2022-11-07)
+
+**Note:** Version bump only for package docs
+
+
+
+
+
+## [0.6.25](https://github.com/cornerstonejs/cornerstone3D-beta/compare/docs@0.6.24...docs@0.6.25) (2022-11-04)
+
+**Note:** Version bump only for package docs
+
+
+
+
+
+## [0.6.24](https://github.com/cornerstonejs/cornerstone3D-beta/compare/docs@0.6.23...docs@0.6.24) (2022-11-04)
+
+**Note:** Version bump only for package docs
+
+
+
+
+
+## [0.6.23](https://github.com/cornerstonejs/cornerstone3D-beta/compare/docs@0.6.22...docs@0.6.23) (2022-11-04)
+
+**Note:** Version bump only for package docs
+
+
+
+
+
+## [0.6.22](https://github.com/cornerstonejs/cornerstone3D-beta/compare/docs@0.6.21...docs@0.6.22) (2022-11-01)
+
+**Note:** Version bump only for package docs
+
+
+
+
+
+## [0.6.21](https://github.com/cornerstonejs/cornerstone3D-beta/compare/docs@0.6.20...docs@0.6.21) (2022-11-01)
+
+**Note:** Version bump only for package docs
+
+
+
+
+
+## [0.6.20](https://github.com/cornerstonejs/cornerstone3D-beta/compare/docs@0.6.19...docs@0.6.20) (2022-10-31)
+
+**Note:** Version bump only for package docs
+
+
+
+
+
 ## [0.6.19](https://github.com/cornerstonejs/cornerstone3D-beta/compare/docs@0.6.18...docs@0.6.19) (2022-10-28)
 
 **Note:** Version bump only for package docs